Refactor utils for creating graph
