+++ src/Components/SampleGraph/SampleGraph.types.ts
@@ -1,7 +1,20 @@
 import { IStyle, IStyleFunctionOrObject } from '@fluentui/react';
 import { IExtendedTheme } from '../../Theming/Theme.types';
 
-export interface ISampleGraphProps {
+export interface IGraphNode<T> {
+    /** color of the node */
+    color?: string;
+    /** data bag to attach to the node */
+    data: T;
+    /** icon to show on the node */
+    icon?: string;
+    /** unique id for the node. Must be unique on the graph */
+    id: string;
+    /** label to show for the node */
+    label: string;
+}
+export interface ISampleGraphProps<T> {
+    nodes: IGraphNode<T>[];
     /**
      * Call to provide customized styling that will layer on top of the variant rules.
      */
